Fix bug in simultaneous change of password and replkey
When both the mongodb admin user password and the replication
key used in the replica set are change at the same time, the current
master is removed from the replica set. What happens is that the
replication key on the master is changed (which removes it from the cluster
temporarily) and then the playbook fails when trying to change the admin
user password (because the master is no longer master). This fix change
the order of tasks to change the password before the replication key is
changed.